+++ scalr/resource_scalr_environment.go
@@ -1,6 +1,7 @@
 package scalr
 
 import (
+	"errors"
 	"fmt"
 	"log"
 
@@ -73,20 +74,50 @@ func resourceScalrEnvironment() *schema.Resource {
 	}
 }
 
+func parseCloudCredentialDefinitions(d *schema.ResourceData) ([]*scalr.CloudCredential, error) {
+	var cloudCredentials []*scalr.CloudCredential
+
+	cloudCredIds := d.Get("cloud_credentials").([]interface{})
+	err := ValidateIDsDefinitions(cloudCredIds)
+	if err != nil {
+		return nil, fmt.Errorf("Got error during parsing cloud credentials: %s", err.Error())
+	}
+
+	for _, cloudCredID := range cloudCredIds {
+		cloudCredentials = append(cloudCredentials, &scalr.CloudCredential{ID: cloudCredID.(string)})
+	}
+
+	return cloudCredentials, nil
+}
+
+func parsePolicyGroupDefinitions(d *schema.ResourceData) ([]*scalr.PolicyGroup, error) {
+	var policyGroups []*scalr.PolicyGroup
+
+	policyGroupIds := d.Get("policy_groups").([]interface{})
+	err := ValidateIDsDefinitions(policyGroupIds)
+	if err != nil {
+		return nil, fmt.Errorf("Got error during parsing policy groups: %s", err.Error())
+	}
+
+	for _, policyGroupID := range policyGroupIds {
+		policyGroups = append(policyGroups, &scalr.PolicyGroup{ID: policyGroupID.(string)})
+	}
+
+	return policyGroups, nil
+}
+
 func resourceScalrEnvironmentCreate(d *schema.ResourceData, meta interface{}) error {
 	scalrClient := meta.(*scalr.Client)
 
 	name := d.Get("name").(string)
 	accountID := d.Get("account_id").(string)
-
-	var cloudCredentials []*scalr.CloudCredential
-	for _, cloudCredsID := range d.Get("cloud_credentials").([]interface{}) {
-		cloudCredentials = append(cloudCredentials, &scalr.CloudCredential{ID: cloudCredsID.(string)})
+	cloudCredentials, err := parseCloudCredentialDefinitions(d)
+	if err != nil {
+		return err
 	}
-
-	var policyGroups []*scalr.PolicyGroup
-	for _, policyGroupID := range d.Get("policy_groups").([]interface{}) {
-		policyGroups = append(policyGroups, &scalr.PolicyGroup{ID: policyGroupID.(string)})
+	policyGroups, err := parsePolicyGroupDefinitions(d)
+	if err != nil {
+		return err
 	}
 
 	options := scalr.EnvironmentCreateOptions{
@@ -114,7 +145,7 @@ func resourceScalrEnvironmentRead(d *schema.ResourceData, meta interface{}) erro
 	log.Printf("[DEBUG] Read configuration of environment: %s", environmentID)
 	environment, err := scalrClient.Environments.Read(ctx, environmentID)
 	if err != nil {
-		if err == scalr.ErrResourceNotFound {
+		if errors.Is(err, scalr.ErrResourceNotFound{}) {
 			// If the resource isn't available, the function should set the ID
 			// to an empty string so Terraform "destroys" the resource in state.
 			d.SetId("")
@@ -161,13 +192,13 @@ func resourceScalrEnvironmentUpdate(d *schema.ResourceData, meta interface{}) er
 	scalrClient := meta.(*scalr.Client)
 
 	var err error
-	var cloudCredentials []*scalr.CloudCredential
-	for _, credsID := range d.Get("cloud_credentials").([]interface{}) {
-		cloudCredentials = append(cloudCredentials, &scalr.CloudCredential{ID: credsID.(string)})
+	cloudCredentials, err := parseCloudCredentialDefinitions(d)
+	if err != nil {
+		return err
 	}
-	var policyGroups []*scalr.PolicyGroup
-	for _, credsID := range d.Get("policy_groups").([]interface{}) {
-		policyGroups = append(policyGroups, &scalr.PolicyGroup{ID: credsID.(string)})
+	policyGroups, err := parsePolicyGroupDefinitions(d)
+	if err != nil {
+		return err
 	}
 
 	// Create a new options struct.
@@ -193,7 +224,7 @@ func resourceScalrEnvironmentDelete(d *schema.ResourceData, meta interface{}) er
 	log.Printf("[DEBUG] Delete environment %s", environmentID)
 	err := scalrClient.Environments.Delete(ctx, d.Id())
 	if err != nil {
-		if err == scalr.ErrResourceNotFound {
+		if errors.Is(err, scalr.ErrResourceNotFound{}) {
 			return nil
 		}
 		return fmt.Errorf(