+++ scalr/resource_scalr_environment_test.go
@@ -2,6 +2,7 @@ package scalr
 
 import (
 	"fmt"
+	"regexp"
 	"testing"
 
 	"github.com/hashicorp/terraform/helper/resource"
@@ -74,6 +75,10 @@ func TestAccEnvironment_update(t *testing.T) {
 					resource.TestCheckResourceAttr("scalr_environment.test", "cost_estimation_enabled", "false"),
 					resource.TestCheckResourceAttr("scalr_environment.test", "cloud_credentials.%", "0"),
 				),
+			},
+			{
+				Config:      testAccEnvironmentUpdateConfigEmptyString(rInt),
+				ExpectError: regexp.MustCompile("Got error during parsing cloud credentials: 0-th value is empty"),
 			},
 		},
 	})
@@ -171,5 +176,15 @@ resource "scalr_environment" "test" {
   account_id = "%s"
   cost_estimation_enabled = false
   cloud_credentials = []
+}`, rInt, defaultAccount)
+}
+
+func testAccEnvironmentUpdateConfigEmptyString(rInt int) string {
+	return fmt.Sprintf(`
+resource "scalr_environment" "test" {
+  name       = "test-env-%d-patched"
+  account_id = "%s"
+  cost_estimation_enabled = false
+  cloud_credentials = [""]
 }`, rInt, defaultAccount)
 }